requirements.txt
google-genai>=0.5.0
Flask>=3.0.0
